simple_preprocessor.py
import os
import pickle
import logging
import numpy as np
import pandas as pd
from datetime import datetime
from collections import defaultdict, Counter
from scipy.stats import entropy, skew, kurtosis
from sklearn.preprocessing import MinMaxScaler, StandardScaler

"""
simple_preprocessor.py
----------------------

This script defines the 'SimpleFeatureEngineeringPreprocessor', which:
    - Maintains a sliding window of the last N packets
    - Computes aggregated/engineered features (e.g. inter-arrival times, rates, stats)
    - Encodes IP/MAC/protocol using label encoders
    - Optionally applies scaling (MinMax or Standard) to numeric features
    - Provides 'add_packet' for incremental usage, and 'transform_packet_list' for batch usage

Key usage:
    1) The detection pipeline (simple_pipeline.py) uses add_packet or transform_packet_list
       to produce a numeric row for LSTM.
    2) This class can be saved/loaded so label encoders and scalers remain consistent.

Important Functions:
    - add_packet(packet_dict):
      * Adds the new packet to the sliding window
      * If the window is not full, engineered features (like interarrival times) are zero
      * Once the window is full, aggregates the entire window to compute stats
      * Encodes, scales, returns a single row of data for that packet

    - transform_packet_list(list_of_dicts):
      * For offline usage (like building a big dataset from a pcap)
      * Processes each packet in sequence, calling add_packet internally

    - save_state() / load_state():
      * Saves/loads label encoders, minmax_scaler, standard_scaler, etc.

How it Connects:
    - 'initial_data_capture.py' uses it to generate CSV with engineered features offline
    - 'simple_pipeline.py' uses it in real-time detection
    - 'pre_processing.py' is a separate script for train vs. transform mode, but this approach
      can handle the same tasks.

To Run:
    - Typically not run directly, but imported into other scripts.
      You can do: from simple_preprocessor import SimpleFeatureEngineeringPreprocessor
      Then instantiate and call add_packet() or transform_packet_list() as needed.
"""

class SimpleFeatureEngineeringPreprocessor:
    def __init__(self, load_state=None, window_size=20, environment=None):
        ...

    def add_packet(self, packet_dict):
        ...

    def _extract_window_features(self):
        ...

    def _zero_engineered_features(self):
        ...

    def _compute_engineered_features(self):
        ...

    def transform_packet_list(self, packets):
        ...

    def _encode_and_scale(self, df):
        ...

    def _encode_categorical_columns(self, df):
        ...

    def _encode_packet_type(self, pkt_type):
        ...

    def _encode_ip(self, ip_str, col):
        ...

    def _encode_mac(self, mac_str, col):
        ...

    def _encode_proto(self, proto_str):
        ...

    def fit_minmax_scaler(self, df):
        ...

    def save_state(self, directory):
        ...

    def load_state(self, directory):
        ...

    def _calculate_interarrival_stats(self, window):
        ...

    def _calculate_packet_size_stats(self, window):
        ...

    def _calculate_tcp_flags_stats(self, window):
        ...

    def _calculate_unique_addr_stats(self, window):
        ...

    def _calculate_arp_anomaly(self, window):
        ...

    def _calculate_error_stats(self, window):
        ...

    def _calculate_modbus_stats(self, window):
        ...

    def _extract_packet_info(self, packet):
        ...