requirements.txt
scapy>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.0.0
tensorflow-macos
tensorflow-metal
shap>=0.41.0
matplotlib>=3.7.0  # Required for SHAP visualizations
seaborn>=0.12.0
python-dotenv==1.0.0
netifaces
flask
flask-cors
openai